Modified the update UI to become more intuitive.
There is now a one-click method of upgrading all packages. When the updater is
first opened, the available updates aren't all selected. This allows the user to
quickly open-close the updater to see if anything interesting is ready to be
updated, without being prompted to 'discard changes'. Clicking the apply-all
button applies all available updates. The user is still able to cherry pick
updates then click 'apply' or 'ok' as previously possible.

svn path=/trunk/playground/sysadmin/kpackagekit/; revision=916804