Fix display of status/role messages in the plasmoid
The QML plasmoid calls a C++ function to get the message strings for the
current status/role. But due to a limitation in Qt4, Enum parameters can
only be used when the Enum is defined in the same class as the C++
function, which is not the case here (the Enums come from PackageKit).
See also https://bugreports.qt.io/browse/QTBUG-19741

So the functions always get passed 0, which means "Unknown", and the
plasmoid always displays "Unknown state"/"Unknown role type".

This patch overloads the necessary functions to accept int parameters as
well to fix this.

BUG: 344639
REVIEW: 123150